// adv_js/final-angular/recipe-saver/src/app/models/recipe.model.ts
export class Recipe {
    id?: string;
    date?: string;
    title?: string;
    link?: string;
    servings?: string;
    time?: string;
    rating?: number;
    image?: string;
    category?: string;
    ingredients?: string;
    instructions?: string;
    notes?: string;
    ingredientsArr?: Array<string>;
}